[refactor] change time unit from `Int` to `Long`

+++ library/ext-functions/src/main/java/ir/kaaveh/ext_functions/Time.kt
@@ -1,9 +1,9 @@
 package ir.kaaveh.ext_functions
 
-val Long.minutes: Long
+val Int.minutes: Int
     get() = this / 60
 
-val Long.seconds: Long
+val Int.seconds: Int
     get() = this % 60
 
 fun String.toTowDigitFormat(): String =
